chore: fix call flow
