+++ scripts/cci-test-timings.data.json
@@ -1,5 +1,5 @@
 {
-  "lastUpdated": "2022-12-30T23:45:39.659Z",
+  "lastUpdated": "2022-12-31T02:03:46.162Z",
   "totalTestFiles": 266,
   "timingData": [
     {
